problem solved, there are two tyes of quotes and that was causing the problem, a copy paste. Labels erased
