+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Cristina <Cristina@student.42.fr>          +#+  +:+       +#+        */
+/*   By: crramire <crramire@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 13:00:59 by crramire          #+#    #+#             */
-/*   Updated: 2024/03/12 07:22:06 by Cristina         ###   ########.fr       */
+/*   Updated: 2024/03/12 12:51:06 by crramire         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,10 +25,10 @@ int	init_data_structure(t_pipex *data)
 	return (NO_ERROR);
 }
 
-void	check_leaks(void)
+/* void	check_leaks(void)
 {
 	system("leaks -q pipex");
-}
+} */
 
 static void	check_program_args(int argc, char **envp)
 {
@@ -48,20 +48,12 @@ int	main(int argc, char **argv, char **envp)
 	int	i = 0;
 
 	//atexit(check_leaks);
-	ft_printf("%s PID: %i\n", argv[0], getpid());
-	while(argv[i] != NULL)
-	{
-		printf("argv[%i] = %s\n", i, argv[i]);
-		i++;
-	}
-	printf("argv[%i] = %s\n", i, argv[i]);
+	//ft_printf("%s PID: %i\n", argv[0], getpid());
 	check_program_args(argc, envp);
 	data.argc = argc;
 	data.argv = argv;
+	i = 0;
 	data.envp = envp;
-	//data.cmds[0] = *(argv + 2);
-	//data.cmds[1] = *(argv + 3);
-	printf("´´´´´´´´´´´´´´´´´´´´´´´´´\nargv[2] = %s\n argv[3] = %s\n", argv[2], argv[3]);
 	if (init_data_structure(&data) == ERROR)
 	{
 		perror("Problem initializing structure");
@@ -72,5 +64,6 @@ int	main(int argc, char **argv, char **envp)
 		perror("No PATH finded");
 		exit(errno);
 	}
+	//REVIEW - He cambiado return por exit
 	return (pipex(&data));
 }
